--- server/api/views.py
@@ -1,38 +1,0 @@
-# api/views.py
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework_simplejwt.tokens import RefreshToken
-from django.contrib.auth.models import User
-from django.contrib.auth import authenticate
-
-
-class LoginView(APIView):
-    def post(self, request):
-        username = request.data.get('username')
-        password = request.data.get('password')
-        user = authenticate(username=username, password=password)
-        if user:
-            refresh = RefreshToken.for_user(user)
-            return Response({
-                'refresh': str(refresh),
-                'access': str(refresh.access_token),
-                'user': {'id': user.id, 'username': user.username, 'email': user.email},
-                'rol': 'Administrador' if user.is_staff else 'Usuario',  # ✅ Añadido
-                'is_staff': user.is_staff,  # ✅ Opcional, pero útil
-            })
-        return Response({'error': 'Credenciales invalidas'}, status=401)
-
-class RegisterView(APIView):
-    def post(self, request):
-        username = request.data.get('username')
-        email = request.data.get('email')
-        password = request.data.get('password')
-        if User.objects.filter(username=username).exists():
-            return Response({'error': 'Error: Este usuario ya existe.'}, status=400)
-        user = User.objects.create_user(username=username, email=email, password=password)
-        refresh = RefreshToken.for_user(user)
-        return Response({
-            'refresh': str(refresh),
-            'access': str(refresh.access_token), 
-            'user': {'id': user.id, 'username': user.username, 'email': user.email}
-        }, status=201)
