+++ server/school/api/alumnos/__init__.py
@@ -1,0 +1,4 @@
+"""
+Módulo de API para Alumnos
+"""
+
