+++ server/school/api/views.py
@@ -5,16 +5,11 @@ from rest_framework.authtoken.models import Token
 from django.contrib.auth import authenticate
 from django.contrib.auth.models import User
 from django.db.models import Q
-from school.models import Periodo, Grado, Materia, Alumno, Calificacion, Personal
+from school.models import Periodo, Grado, Materia, Calificacion, Personal
 from school.api.serializers import (
     PeriodoSerializer, GradoSerializer,
-    MateriaSerializer, AlumnoSerializer,
+    MateriaSerializer,
     CalificacionSerializer, PersonalSerializer, UserSerializer
-)
-from school.exceptions.domain_exceptions import (
-    AlumnoNotFoundError,
-    MatriculaDuplicadaError,
-    GradoNotFoundError
 )
 
 
@@ -50,148 +45,6 @@ class MateriaViewSet(viewsets.ModelViewSet):
             queryset = queryset.filter(profesor_id=profesor_id)
         return queryset
 
-
-class AlumnoViewSet(viewsets.ViewSet):
-    """ViewSet delgado que delega a servicios"""
-    permission_classes = []
-    
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        from school.services.alumno_service import AlumnoService
-        self.alumno_service = AlumnoService()
-    
-    def list(self, request):
-        """Lista alumnos con filtros"""
-        filtros = {
-            'grado_id': request.query_params.get('grado'),
-            'institucion_id': request.query_params.get('institucion'),
-            'search': request.query_params.get('search')
-        }
-        # Limpiar None values
-        filtros = {k: v for k, v in filtros.items() if v is not None}
-        
-        try:
-            alumnos = self.alumno_service.listar_alumnos(filtros)
-            serializer = AlumnoSerializer(alumnos, many=True)
-            return Response(serializer.data)
-        except Exception as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-    
-    def retrieve(self, request, pk=None):
-        """Obtiene un alumno por ID"""
-        try:
-            alumno = self.alumno_service.obtener_alumno(int(pk))
-            serializer = AlumnoSerializer(alumno)
-            return Response(serializer.data)
-        except AlumnoNotFoundError as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        except ValueError:
-            return Response(
-                {'error': 'ID inválido'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-    
-    def create(self, request):
-        """Crea un nuevo alumno"""
-        serializer = AlumnoSerializer(data=request.data)
-        if not serializer.is_valid():
-            return Response(
-                serializer.errors,
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        
-        try:
-            datos_validados = serializer.validated_data
-            # Mapear campos del serializer a campos del servicio
-            datos_servicio = {
-                'nombre': datos_validados.get('nombre'),
-                'apellido': datos_validados.get('apellido'),
-                'matricula': datos_validados.get('matricula'),
-                'fecha_nacimiento': datos_validados.get('fecha_nacimiento') or datos_validados.get('fechaNacimiento'),
-                'correo': datos_validados.get('correo') or datos_validados.get('email'),
-                'grado_id': datos_validados.get('grado') or datos_validados.get('gradoEstudioId')
-            }
-            
-            alumno = self.alumno_service.crear_alumno(datos_servicio)
-            return Response(
-                AlumnoSerializer(alumno).data,
-                status=status.HTTP_201_CREATED
-            )
-        except (MatriculaDuplicadaError, GradoNotFoundError) as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        except Exception as e:
-            return Response(
-                {'error': f'Error al crear alumno: {str(e)}'},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-    
-    def update(self, request, pk=None):
-        """Actualiza un alumno"""
-        serializer = AlumnoSerializer(data=request.data)
-        if not serializer.is_valid():
-            return Response(
-                serializer.errors,
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        
-        try:
-            datos_validados = serializer.validated_data
-            datos_servicio = {
-                'nombre': datos_validados.get('nombre'),
-                'apellido': datos_validados.get('apellido'),
-                'matricula': datos_validados.get('matricula'),
-                'fecha_nacimiento': datos_validados.get('fecha_nacimiento') or datos_validados.get('fechaNacimiento'),
-                'correo': datos_validados.get('correo') or datos_validados.get('email'),
-                'grado_id': datos_validados.get('grado') or datos_validados.get('gradoEstudioId')
-            }
-            
-            alumno = self.alumno_service.actualizar_alumno(int(pk), datos_servicio)
-            return Response(AlumnoSerializer(alumno).data)
-        except AlumnoNotFoundError as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        except (MatriculaDuplicadaError, GradoNotFoundError) as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        except ValueError:
-            return Response(
-                {'error': 'ID inválido'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        except Exception as e:
-            return Response(
-                {'error': f'Error al actualizar alumno: {str(e)}'},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-    
-    def destroy(self, request, pk=None):
-        """Elimina un alumno"""
-        try:
-            self.alumno_service.eliminar_alumno(int(pk))
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except AlumnoNotFoundError as e:
-            return Response(
-                {'error': str(e)},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        except ValueError:
-            return Response(
-                {'error': 'ID inválido'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
 
 
 class CalificacionViewSet(viewsets.ModelViewSet):
@@ -239,31 +92,79 @@ class PersonalViewSet(viewsets.ModelViewSet):
         return queryset
 
 
-# Vista para autenticación
+# Authentication ViewSet for user login/logout operations
 class AuthViewSet(viewsets.ViewSet):
+    """
+    ViewSet for handling user authentication.
+    
+    Endpoints:
+        POST /api/auth/login/ - User login
+        POST /api/auth/logout/ - User logout
+        GET /api/auth/me/ - Get current user info
+        
+    Educational System Roles:
+        - Administrador: is_superuser = True (full system access)
+        - Docente: is_staff = True (teacher access to courses/students)
+        - Estudiante: regular user (student access to learning materials)
+    """
     permission_classes = []
     
     @action(detail=False, methods=['post'])
     def login(self, request):
+        """
+        Authenticate user and return token with user data.
+        
+        IMPORTANT FIX: The 'rol' field must be included inside the 'user' object
+        to prevent frontend error: "Cannot read properties of undefined (reading 'rol')"
+        
+        The frontend (auth.ts) expects: response.user.rol
+        
+        Request Body:
+            - username (str): User's username
+            - password (str): User's password
+            
+        Returns:
+            200 OK: { token, user: { id, username, email, rol, ... } }
+            401 Unauthorized: { error: "Invalid credentials" }
+        """
         username = request.data.get('username')
         password = request.data.get('password')
         
         if not username or not password:
             return Response(
-                {'error': 'Username y password son requeridos'}, 
+                {'error': 'Username and password are required'}, 
                 status=status.HTTP_400_BAD_REQUEST
             )
         
         user = authenticate(username=username, password=password)
         if user:
-            token, created = Token.objects.get_or_create(user=user)
+            token, _ = Token.objects.get_or_create(user=user)
+            
+            # Determine user role based on Django permissions:
+            # - Administrador: superuser (full access to system management)
+            # - Docente: staff member (can manage courses and students)
+            # - Estudiante: regular user (can access learning materials)
+            if user.is_superuser:
+                rol = 'Administrador'
+            elif user.is_staff:
+                rol = 'Docente'
+            else:
+                rol = 'Estudiante'
+            
+            # Include 'rol' in user object for frontend compatibility (auth.ts expects response.user.rol)
+            user_data = UserSerializer(user).data
+            user_data['rol'] = rol  # Add role to user object
+            
             return Response({
                 'token': token.key,
-                'user': UserSerializer(user).data
+                'user': user_data,
+                'rol': rol,  # Also at top level for flexibility
+                'is_staff': user.is_staff,
+                'is_superuser': user.is_superuser,
             })
         else:
             return Response(
-                {'error': 'Credenciales inválidas'}, 
+                {'error': 'Invalid credentials'}, 
                 status=status.HTTP_401_UNAUTHORIZED
             )
     
@@ -273,9 +174,9 @@ class AuthViewSet(viewsets.ViewSet):
             try:
                 request.user.auth_token.delete()
                 return Response({'message': 'Logout exitoso'})
-            except:
+            except Exception:
                 return Response(
-                    {'error': 'Error al hacer logout'}, 
+                    {'error': 'Error during logout'}, 
                     status=status.HTTP_400_BAD_REQUEST
                 )
         return Response(
